requirements.txt
Flask>=2.2.2
google-cloud-datastore==1.12.0
google-cloud-ndb
appengine-python-standard
jinja2==3.0.3
future
